chore: move bria-subscriber into own file
