+++ src/services/bria/errors.ts
@@ -1,0 +1,3 @@
+import { DomainError } from "@domain/shared"
+
+export class BriaEventError extends DomainError {}
